backend/requirements.txt
# Core
fastapi
uvicorn
python-dotenv

# Auth
python-jose[cryptography]
passlib[bcrypt]
python-multipart

# Database
sqlalchemy
alembic
psycopg2-binary

# GitHub Integration
PyGithub

# Slack Integration
slack-sdk

# HTTP Client
httpx
aiohttp

# Pydantic
pydantic
pydantic-settings
email-validator

# Utilities
python-slugify
python-decouple
python-dateutil
pyjwt
cryptography

# Testing
pytest
pytest-cov
pytest-asyncio>=0.21.0,<0.23.0
pytest-mock>=3.10.0,<4.0.0
responses>=0.23.0,<0.24.0

# Development
black>=23.0.0,<24.0.0
isort>=5.12.0,<6.0.0
mypy>=1.0.0,<2.0.0
flake8>=6.0.0,<7.0.0
pre-commit>=3.0.0,<4.0.0

# Documentation
mkdocs>=1.4.0,<2.0.0
mkdocs-material>=9.0.0,<10.0.0
mkdocstrings[python]>=0.20.0,<0.22.0

# Monitoring and Logging
sentry-sdk>=1.20.0,<2.0.0
structlog>=23.0.0,<24.0.0
